Use Phaser timelines for fast break
