+++ FrontEnd/static/js/phaser/animation/turnAnimation.js
@@ -9,9 +9,9 @@ import {
   animatePutbackAttempt,
   animateKickoutReset
 } from "./ballManager.js";
-import { tweenBallTo, runPass, PASS_DEBUG, tweenPlayerTo, detachBall } from "./ballTween.js";
+import { tweenBallTo, runPass, PASS_DEBUG, tweenPlayerTo } from "./ballTween.js";
 import animationConfig from "./animation_config.js";
-import { HOME_RIM_COORDS, AWAY_RIM_COORDS, HOME_TOP_KEY, AWAY_TOP_KEY } from "./courtConstants.js";
+import { HOME_RIM_COORDS, AWAY_RIM_COORDS } from "./courtConstants.js";
 import { DEBUG } from "../utils/debug.js";
 import { States } from "../state/gameStateMachine.js";
 import {
@@ -615,245 +615,6 @@ async function runInboundSetup({
   scene.isInboundSetup = false;
 }
 
-async function runFastBreakSequence({ scene, turnData, playerSprites, ballSprite }) {
-  if (!scene || !turnData || scene.skipToEnd) return;
-  if (!scene.ballSprite) scene.ballSprite = ballSprite;
-
-  scene.stateMachine?.transition(States.FastBreak);
-  scene.events?.emit("fb:start");
-
-  if (scene.tweens) {
-    for (const sprite of Object.values(playerSprites)) {
-      scene.tweens.killTweensOf(sprite);
-    }
-  }
-
-  try {
-    const animations = turnData.animations || [];
-    const width = scene.game.config.width;
-    const height = scene.game.config.height;
-    const sprintDuration =
-      animationConfig.fastBreak?.sprintDuration ?? 800;
-
-    const ownerAnim = animations.find(a => a.hasBallAtStep?.[0]);
-    if (ownerAnim) {
-      const ownerSprite = playerSprites[ownerAnim.playerId];
-      if (ownerSprite) attachBallToPlayer(scene, ballSprite, ownerSprite);
-    }
-
-    const sprintPromises = [];
-    for (const anim of animations) {
-      const sprite = playerSprites[anim.playerId];
-      if (!sprite) continue;
-      const start = anim.start || anim.movement?.[0]?.coords;
-      const end = anim.end || anim.movement?.[anim.movement.length - 1]?.coords;
-      if (!start || !end) continue;
-      const startPx = gridToPixels(start.x, start.y, width, height);
-      const endPx = gridToPixels(end.x, end.y, width, height);
-      sprite.setPosition(startPx.x, startPx.y);
-      sprintPromises.push(
-        tweenPlayerTo(scene, sprite, endPx, {
-          duration: sprintDuration,
-          easing: "Sine.easeInOut"
-        })
-      );
-    }
-    await Promise.all(sprintPromises);
-    if (scene.skipToEnd) return;
-
-    const passes = turnData.passes || [];
-    if (passes.length > 0) {
-      for (const p of passes) {
-        if (scene.skipToEnd) break;
-        await runPass(scene, {
-          fromId: p.fromId || p.from_id,
-          toId: p.toId || p.to_id,
-          duration: p.duration || animationConfig.pass.duration,
-          easing: animationConfig.pass.easing
-        });
-      }
-    } else {
-      const passEvents = [];
-      for (const anim of animations) {
-        const moves = anim.movement || [];
-        for (const step of moves) {
-          if (step.action === "pass") {
-            const ts = step.timestamp;
-            const receiverAnim = animations.find(a =>
-              a.movement?.some(m => m.action === "receive" && m.timestamp === ts)
-            );
-            const receiveStep = receiverAnim?.movement.find(
-              m => m.action === "receive" && m.timestamp === ts
-            );
-            const delta = receiveStep
-              ? receiveStep.timestamp - ts
-              : animationConfig.pass.duration;
-            passEvents.push({
-              timestamp: ts,
-              fromId: anim.playerId,
-              toId: receiverAnim?.playerId,
-              duration: delta
-            });
-          }
-        }
-      }
-      passEvents.sort((a, b) => a.timestamp - b.timestamp);
-      for (const evt of passEvents) {
-        if (scene.skipToEnd) break;
-        await runPass(scene, {
-          fromId: evt.fromId,
-          toId: evt.toId,
-          duration: evt.duration,
-          easing: animationConfig.pass.easing
-        });
-      }
-    }
-    if (scene.skipToEnd) return;
-
-    if (turnData.hold_up) {
-      const bhId = getCurrentOwner(scene);
-      const bhSprite = bhId != null ? playerSprites[bhId] : null;
-      const stopperId = turnData.stopper_id;
-      const stopperSprite = stopperId != null ? playerSprites[stopperId] : null;
-      const promises = [];
-
-      if (bhSprite) {
-        attachBallToPlayer(scene, ballSprite, bhSprite);
-        const topKey = bhSprite.team === "home" ? HOME_TOP_KEY : AWAY_TOP_KEY;
-        const topKeyPx = gridToPixels(topKey.x, topKey.y, width, height);
-        promises.push(
-          tweenPlayerTo(scene, bhSprite, topKeyPx, {
-            duration: sprintDuration / 2,
-            easing: "Sine.easeInOut"
-          })
-        );
-
-        const rim = bhSprite.team === "home" ? HOME_RIM_COORDS : AWAY_RIM_COORDS;
-
-        if (stopperSprite) {
-          const offsetX = bhSprite.team === "home" ? 6 : -6;
-          const stopGrid = {
-            x: topKey.x + offsetX,
-            y: topKey.y + Phaser.Math.Between(-3, 3)
-          };
-          const stopPx = gridToPixels(stopGrid.x, stopGrid.y, width, height);
-          promises.push(
-            tweenPlayerTo(scene, stopperSprite, stopPx, {
-              duration: sprintDuration / 2,
-              easing: "Sine.easeInOut"
-            })
-          );
-        }
-
-        const inPlayDef = new Set(
-          (turnData.roles?.defense || []).map(id => String(id))
-        );
-        for (const [id, sprite] of Object.entries(playerSprites)) {
-          if (sprite === bhSprite || sprite === stopperSprite) continue;
-          if (sprite.team === bhSprite.team) {
-            const spotGrid = {
-              x: Phaser.Math.Between(50, 51),
-              y: Phaser.Math.Between(10, 40)
-            };
-            const spotPx = gridToPixels(spotGrid.x, spotGrid.y, width, height);
-            promises.push(
-              tweenPlayerTo(scene, sprite, spotPx, {
-                duration: sprintDuration / 2,
-                easing: "Sine.easeInOut"
-              })
-            );
-          } else {
-            if (inPlayDef.has(String(id))) {
-              const laneGrid = {
-                x: Phaser.Math.Between(
-                  Math.min(topKey.x, rim.x) + 1,
-                  Math.max(topKey.x, rim.x) - 1
-                ),
-                y: Phaser.Math.Between(10, 40)
-              };
-              const lanePx = gridToPixels(laneGrid.x, laneGrid.y, width, height);
-              promises.push(
-                tweenPlayerTo(scene, sprite, lanePx, {
-                  duration: sprintDuration / 2,
-                  easing: "Sine.easeInOut"
-                })
-              );
-            } else {
-              const halfGrid = {
-                x: Phaser.Math.Between(50, 51),
-                y: Phaser.Math.Between(10, 40)
-              };
-              const halfPx = gridToPixels(halfGrid.x, halfGrid.y, width, height);
-              promises.push(
-                tweenPlayerTo(scene, sprite, halfPx, {
-                  duration: sprintDuration / 2,
-                  easing: "Sine.easeInOut"
-                })
-              );
-            }
-          }
-        }
-      }
-
-      await Promise.all(promises);
-      if (typeof scene.startNextHalfCourtOffense === "function") {
-        scene.startNextHalfCourtOffense();
-      }
-      return;
-    }
-
-    const shooterId =
-      turnData.shooterId || turnData.shooter_id || getCurrentOwner(scene);
-    const shooterSprite = shooterId != null ? playerSprites[shooterId] : null;
-    if (shooterSprite) {
-      attachBallToPlayer(scene, ballSprite, shooterSprite);
-      const rimGrid =
-        shooterSprite.team === "home" ? HOME_RIM_COORDS : AWAY_RIM_COORDS;
-      const rimPx = gridToPixels(rimGrid.x, rimGrid.y, width, height);
-      await tweenPlayerTo(scene, shooterSprite, rimPx, {
-        duration: sprintDuration / 2,
-        easing: "Sine.easeInOut"
-      });
-      detachBall(scene, ballSprite);
-      const arcHeight = animationConfig.fastBreak?.arcHeight ?? 50;
-      await tweenBallTo(scene, ballSprite, rimPx, {
-        duration: animationConfig.inbound.duration,
-        easing: animationConfig.inbound.easing,
-        arc: { height: arcHeight }
-      });
-      if (turnData.result_type === "MAKE") {
-        const rimHoldMs = animationConfig.fastBreak?.rimHoldMs ?? 2000;
-        await new Promise((resolve) =>
-          scene.time.delayedCall(rimHoldMs, resolve)
-        );
-        const newOffenseSide =
-          shooterSprite.team === "home" ? "away" : "home";
-        await runInboundSetup({
-          scene,
-          ballSprite,
-          playerSprites,
-          newOffenseSide
-        });
-      } else {
-        await animateRebound({
-          scene,
-          ballSprite,
-          playerSprites,
-          animations,
-          rebounderId:
-            turnData.rebounder_player_id ||
-            turnData.rebounderId ||
-            turnData.rebounder_id,
-          ballSpot: turnData.ballSpot || turnData.ball_spot,
-          shooterId: turnData.shooterId || turnData.shooter_id
-        });
-      }
-    }
-  } finally {
-    if (scene.stateMachine?.is(States.FastBreak)) scene.stateMachine?.transition(States.HalfCourt);
-    scene.events?.emit("fb:end");
-  }
-}
 
 /**
  * Step-synchronized possession animation.
@@ -1190,7 +951,7 @@ export async function playTurnAnimation({ scene, simData, playerSprites, turnDat
   }
 }
 
-export { runInboundSetup, runSideInboundSetup, runFastBreakSequence, runDefensiveReboundSetup };
+export { runInboundSetup, runSideInboundSetup, runDefensiveReboundSetup };
 
 if (typeof window !== "undefined") {
   window.playTurnAnimation = playTurnAnimation;
